+++ CHANGELOG.md
@@ -5,6 +5,10 @@ Formatted as described on http://keepachangelog.com/.
 
 ## Unreleased
 
+### Fixes
+
+- Fetch fragment with no molblock throws error (#41)
+
 ## [2.1.0] - 2017-01-17
 
 ### Added